cleanups from #2532
