+++ apps/devmo/helpers.py
@@ -1,5 +1,4 @@
 import datetime
-from polib import pofile
 import re
 import urllib
 from urlobject import URLObject
@@ -30,19 +29,6 @@ register.filter(strip_tags)
 register.filter(defaultfilters.timesince)
 register.filter(defaultfilters.truncatewords)
 register.filter(entity_decode)
-
-
-def strings_are_translated(strings, locale):
-    # http://stackoverflow.com/a/24339946/571420
-    po = pofile(os.path.join(settings.ROOT,
-                             'locale/%s/LC_MESSAGES/messages.po' % locale))
-    all_strings_translated = True
-    for string in strings:
-        if not any(e for e in po if e.msgid == string and
-                   (e.translated() and 'fuzzy' not in e.flags)
-                   and not e.obsolete):
-            all_strings_translated = False
-    return all_strings_translated
 
 
 @register.function
