+++ kuma/users/views.py
@@ -29,7 +29,7 @@ from taggit.utils import parse_tags
 from teamwork.models import Team
 from waffle import switch_is_active
 
-from devmo.helpers import strings_are_translated
+from devmo.utils import strings_are_translated
 from demos.models import Submission
 from sumo.decorators import ssl_required
 from sumo.urlresolvers import reverse, split_path
@@ -207,11 +207,9 @@ def browserid_register(request):
                 auth.login(request, user)
 
                 if switch_is_active('welcome_email'):
-                    if request.locale == settings.WIKI_DEFAULT_LANGUAGE:
-                            send_welcome_email.delay(user.pk)
-                    else:
-                        if strings_are_translated(WELCOME_EMAIL_STRINGS,
-                                                  request.locale):
+                    if (request.locale == settings.WIKI_DEFAULT_LANGUAGE or
+                        strings_are_translated(WELCOME_EMAIL_STRINGS,
+                                               request.locale)):
                             send_welcome_email.delay(user.pk)
 
                 newsletter_subscribe(request, email,
